move init_defaults to tempo-node crate
Move init_download_urls, init_payload_builder_defaults, init_txpool_defaults
and the public init_defaults function from bin/tempo into crates/node/src/defaults.rs.

Re-export init_defaults from tempo_node lib.rs alongside init_version_metadata.

Telemetry CLI types (TelemetryArgs, UrlWithAuth, TelemetryConfig) remain in
the tempo binary as they are CLI-specific.

Amp-Thread-ID: https://ampcode.com/threads/T-019c76ac-cb19-70fd-aa37-75be4e69ea25
