+++ bin/tempo/src/defaults.rs
@@ -1,13 +1,8 @@
 use base64::{Engine, prelude::BASE64_STANDARD};
 use eyre::Context as _;
 use jiff::SignedDuration;
-use reth_cli_commands::download::DownloadDefaults;
-use reth_ethereum::node::core::args::{DefaultPayloadBuilderValues, DefaultTxPoolValues};
-use std::{borrow::Cow, str::FromStr, time::Duration};
-use tempo_chainspec::hardfork::TempoHardfork;
+use std::str::FromStr;
 use url::Url;
-
-pub(crate) const DEFAULT_DOWNLOAD_URL: &str = "https://snapshots.tempoxyz.dev/4217";
 
 /// Default OTLP logs filter level for telemetry.
 const DEFAULT_LOGS_OTLP_FILTER: &str = "debug";
@@ -118,59 +113,4 @@ pub(crate) struct TelemetryConfig {
     pub(crate) metrics_prometheus_interval: SignedDuration,
     /// Authorization header for metrics push
     pub(crate) metrics_auth_header: Option<String>,
-}
-
-fn init_download_urls() {
-    let download_defaults = DownloadDefaults {
-        available_snapshots: vec![
-            Cow::Owned(format!("{DEFAULT_DOWNLOAD_URL} (mainnet)")),
-            Cow::Borrowed("https://snapshots.tempoxyz.dev/42431 (moderato)"),
-            Cow::Borrowed("https://snapshots.tempoxyz.dev/42429 (andantino)"),
-        ],
-        default_base_url: Cow::Borrowed(DEFAULT_DOWNLOAD_URL),
-        default_chain_aware_base_url: None,
-        long_help: None,
-    };
-
-    download_defaults
-        .try_init()
-        .expect("failed to initialize download URLs");
-}
-
-fn init_payload_builder_defaults() {
-    DefaultPayloadBuilderValues::default()
-        .with_interval(Duration::from_millis(100))
-        .with_max_payload_tasks(16)
-        .with_deadline(4)
-        .try_init()
-        .expect("failed to initialize payload builder defaults");
-}
-
-fn init_txpool_defaults() {
-    DefaultTxPoolValues::default()
-        .with_pending_max_count(50000)
-        .with_basefee_max_count(50000)
-        .with_queued_max_count(50000)
-        .with_pending_max_size(100)
-        .with_basefee_max_size(100)
-        .with_queued_max_size(100)
-        .with_no_locals(true)
-        .with_max_queued_lifetime(Duration::from_secs(120))
-        .with_max_new_pending_txs_notifications(150000)
-        .with_max_account_slots(150000)
-        .with_pending_tx_listener_buffer_size(50000)
-        .with_new_tx_listener_buffer_size(50000)
-        .with_disable_transactions_backup(true)
-        .with_additional_validation_tasks(8)
-        .with_minimal_protocol_basefee(TempoHardfork::default().base_fee())
-        .with_minimum_priority_fee(Some(0))
-        .with_max_batch_size(50000)
-        .try_init()
-        .expect("failed to initialize txpool defaults");
-}
-
-pub(crate) fn init_defaults() {
-    init_download_urls();
-    init_payload_builder_defaults();
-    init_txpool_defaults();
 }
