+++ crates/node/Cargo.toml
@@ -21,6 +21,8 @@ tempo-payload-types.workspace = true
 tempo-precompiles.workspace = true
 tempo-primitives = { workspace = true, features = ["default"] }
 
+base64.workspace = true
+reth-cli-commands.workspace = true
 thiserror.workspace = true
 
 reth-errors.workspace = true
